pyproject.toml
[tool.poetry]
name = "gpu-text-harvest"
version = "0.1.0"
description = "GPU-accelerated batch PDF text extraction wrapper for marker-pdf on NVIDIA GraceBlackwell"
authors = ["Craig Trim <craigtrim@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
marker-pdf = ">=1.0.0"
torch = ">=2.0.0"
psutil = ">=7.1.3,<8.0.0"

[tool.poetry.scripts]
harvest = "gpu_text_harvest.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
